Moved get account into its own controller
