+++ conf/v1.routes
@@ -8,7 +8,7 @@ GET       /admin/metrics                                     @com.kenshoo.play.m
 GET       /manager/:id                                                              @uk.gov.hmrc.lisaapi.controllers.DiscoverController.discover(id: String)
 POST      /manager/:id/investors                                                    @uk.gov.hmrc.lisaapi.controllers.InvestorController.createLisaInvestor(id: String)
 POST      /manager/:id/accounts                                                     @uk.gov.hmrc.lisaapi.controllers.AccountController.createOrTransferLisaAccount(id: String)
-GET       /manager/:id/accounts/:accountId                                          @uk.gov.hmrc.lisaapi.controllers.AccountController.getAccountDetails(id: String, accountId: String)
+GET       /manager/:id/accounts/:accountId                                          @uk.gov.hmrc.lisaapi.controllers.GetAccountController.getAccountDetails(id: String, accountId: String)
 POST      /manager/:id/accounts/:accountId/close-account                            @uk.gov.hmrc.lisaapi.controllers.CloseAccountController.closeLisaAccount(id: String, accountId: String)
 POST      /manager/:id/accounts/reinstate-account                                   @uk.gov.hmrc.lisaapi.controllers.ReinstateAccountController.reinstateAccount(id: String)
 POST      /manager/:id/accounts/:accountId/update-subscription                      @uk.gov.hmrc.lisaapi.controllers.UpdateSubscriptionController.updateSubscription(id: String, accountId: String)
