+++ test/unit/controllers/AccountControllerSpec.scala
@@ -16,7 +16,6 @@
 
 package unit.controllers
 
-import org.joda.time.DateTime
 import org.mockito.Matchers.{eq => matchersEquals, _}
 import org.mockito.Mockito._
 import org.scalatest.BeforeAndAfterEach
@@ -703,157 +702,6 @@ class AccountControllerSpec extends PlaySpec with MockitoSugar with OneAppPerSui
 
   }
 
-  "The Get Account Details endpoint" must {
-
-    when(mockAuthCon.authorise[Option[String]](any(), any())(any(), any())).thenReturn(Future(Some("1234")))
-
-    "return the correct json" when {
-      "returning a valid open account response" in {
-        when(mockService.getAccount(any(), any())(any())).thenReturn(Future.successful(
-          GetLisaAccountSuccessResponse(
-            investorId = "9876543210",
-            accountId = "8765432100",
-            creationReason = "New",
-            firstSubscriptionDate = new DateTime(validDate),
-            accountStatus = "OPEN",
-            subscriptionStatus = "AVAILABLE",
-            accountClosureReason = None,
-            closureDate = None,
-            transferAccount = None
-          )
-        ))
-
-        doSyncGetAccountDetailsRequest(res => {
-          status(res) mustBe OK
-          val json = contentAsJson(res)
-
-          (json \ "investorId").as[String] mustBe "9876543210"
-          (json \ "accountId").as[String] mustBe "8765432100"
-          (json \ "creationReason").as[String] mustBe "New"
-          (json \ "firstSubscriptionDate").as[String] mustBe validDate
-          (json \ "accountStatus").as[String] mustBe "OPEN"
-          (json \ "subscriptionStatus").as[String] mustBe "AVAILABLE"
-          (json \ "accountClosureReason").asOpt[String] mustBe None
-          (json \ "closureDate").asOpt[String] mustBe None
-          (json \ "transferAccount").asOpt[JsObject] mustBe None
-        })
-      }
-      "returning a valid close account response" in {
-        when(mockService.getAccount(any(), any())(any())).thenReturn(Future.successful(
-          GetLisaAccountSuccessResponse(
-            investorId = "9876543210",
-            accountId = "8765432100",
-            creationReason = "New",
-            firstSubscriptionDate = new DateTime(validDate),
-            accountStatus = "CLOSED",
-            subscriptionStatus = "ACTIVE",
-            accountClosureReason = Some("All funds withdrawn"),
-            closureDate = Some(new DateTime(validDate)),
-            transferAccount = None)
-        ))
-
-        doSyncGetAccountDetailsRequest(res => {
-          status(res) mustBe OK
-
-          val json = contentAsJson(res)
-
-          (json \ "investorId").as[String] mustBe "9876543210"
-          (json \ "accountId").as[String] mustBe "8765432100"
-          (json \ "creationReason").as[String] mustBe "New"
-          (json \ "firstSubscriptionDate").as[String] mustBe validDate
-          (json \ "accountStatus").as[String] mustBe "CLOSED"
-          (json \ "subscriptionStatus").as[String] mustBe "ACTIVE"
-          (json \ "accountClosureReason").asOpt[String] mustBe Some("All funds withdrawn")
-          (json \ "closureDate").asOpt[String] mustBe Some(validDate)
-          (json \ "transferAccount").asOpt[JsObject] mustBe None
-        })
-      }
-      "returning a valid transfer account response" in {
-        when(mockService.getAccount(any(), any())(any())).thenReturn(Future.successful(
-          GetLisaAccountSuccessResponse(
-            investorId = "9876543210",
-            accountId = "8765432100",
-            creationReason = "Transferred",
-            firstSubscriptionDate = new DateTime(validDate),
-            accountStatus = "OPEN",
-            subscriptionStatus = "ACTIVE",
-            accountClosureReason = None,
-            closureDate = None,
-            transferAccount = Some(
-              GetLisaAccountTransferAccount(
-                "8765432102",
-                "Z543333",
-                new DateTime(validDate)
-              )
-            )
-          )
-        ))
-
-        doSyncGetAccountDetailsRequest(res => {
-          status(res) mustBe OK
-
-          val json = contentAsJson(res)
-
-          (json \ "investorId").as[String] mustBe "9876543210"
-          (json \ "accountId").as[String] mustBe "8765432100"
-          (json \ "creationReason").as[String] mustBe "Transferred"
-          (json \ "firstSubscriptionDate").as[String] mustBe validDate
-          (json \ "accountStatus").as[String] mustBe "OPEN"
-          (json \ "subscriptionStatus").as[String] mustBe "ACTIVE"
-          (json \ "accountClosureReason").asOpt[String] mustBe None
-          (json \ "closureDate").asOpt[String] mustBe None
-          (json \ "transferAccount" \ "transferredFromAccountId").as[String] mustBe "8765432102"
-          (json \ "transferAccount" \ "transferredFromLMRN").as[String] mustBe "Z543333"
-          (json \ "transferAccount" \ "transferInDate").as[String] mustBe validDate
-        })
-      }
-      "returning a valid void account response" in {
-        when(mockService.getAccount(any(), any())(any())).thenReturn(Future.successful(
-          GetLisaAccountSuccessResponse(
-            investorId = "9876543210",
-            accountId = "8765432100",
-            creationReason = "New",
-            firstSubscriptionDate = new DateTime(validDate),
-            accountStatus = "VOID",
-            subscriptionStatus = "ACTIVE",
-            accountClosureReason = None,
-            closureDate = None,
-            transferAccount = None
-          )
-        ))
-
-        doSyncGetAccountDetailsRequest(res => {
-          status(res) mustBe OK
-
-          val json = contentAsJson(res)
-
-          (json \ "investorId").as[String] mustBe "9876543210"
-          (json \ "accountId").as[String] mustBe "8765432100"
-          (json \ "creationReason").as[String] mustBe "New"
-          (json \ "firstSubscriptionDate").as[String] mustBe validDate
-          (json \ "accountStatus").as[String] mustBe "VOID"
-          (json \ "subscriptionStatus").as[String] mustBe "ACTIVE"
-          (json \ "accountClosureReason").asOpt[String] mustBe None
-          (json \ "closureDate").asOpt[String] mustBe None
-          (json \ "transferAccount").asOpt[JsObject] mustBe None
-        })
-      }
-      "returning a account not found response" in {
-        when(mockService.getAccount(any(), any())(any())).thenReturn(Future.successful(GetLisaAccountDoesNotExistResponse))
-        doSyncGetAccountDetailsRequest(res => {
-          (contentAsJson(res) \ "code").as[String] mustBe "INVESTOR_ACCOUNTID_NOT_FOUND"
-        })
-      }
-      "returning a internal server error response" in {
-        when(mockService.getAccount(any(), any())(any())).thenReturn(Future.successful(GetLisaAccountErrorResponse))
-        doSyncGetAccountDetailsRequest(res => {
-          (contentAsJson(res) \ "code").as[String] mustBe "INTERNAL_SERVER_ERROR"
-        })
-      }
-    }
-
-  }
-
   def doCreateOrTransferRequest(jsonString: String, lmrn: String = lisaManager)(callback: (Future[Result]) => Unit) {
     val res = SUT.createOrTransferLisaAccount(lmrn).apply(FakeRequest(Helpers.PUT, "/").withHeaders(acceptHeader).
       withBody(AnyContentAsJson(Json.parse(jsonString))))
@@ -865,11 +713,6 @@ class AccountControllerSpec extends PlaySpec with MockitoSugar with OneAppPerSui
     val res = await(SUT.createOrTransferLisaAccount(lisaManager).apply(FakeRequest(Helpers.PUT, "/").withHeaders(acceptHeader).
       withBody(AnyContentAsJson(Json.parse(jsonString)))))
 
-    callback(res)
-  }
-
-  def doSyncGetAccountDetailsRequest(callback: (Future[Result]) => Unit) {
-    val res = SUT.getAccountDetails(lisaManager, accountId).apply(FakeRequest(Helpers.GET, "/").withHeaders(acceptHeader))
     callback(res)
   }
